+++ dee/CHANGELOG.md
@@ -6,10 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.0.12] - 2023-08-09
+
 ### Added
 
 - dee: add metadata for future beacons
-- dee: inspect tlock age header
+- dee: add inspect for tlock age header
 
 ## [0.0.11] - 2023-08-08